requirements.txt
python-dotenv
PyGithub~=1.55